requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dotenv==0.14.0
python-json-logger==0.1.11
Werkzeug==1.0.1
